pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "anthropic>=0.40",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "slack-sdk>=3.33",
    "yfinance>=0.2",
    "jinja2>=3.1",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "pyyaml>=6.0",
    "dropbox>=12.0",
    "python-multipart>=0.0.12",
    "aiosmtplib>=3.0",
    "aiohttp>=3.13.3",
    "icalendar>=7.0.3",
    "recurring-ical-events>=3.8.1",
    "caldav>=3.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app", "jobs"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]
